requirements.txt
langchain-core>=0.1.0
langchain-ollama>=0.0.3
langchain-community>=0.0.13
langchain-openrouter>=0.0.1
python-dotenv>=1.0.0
typing-extensions>=4.8.0
pydantic>=2.0.0
aiohttp>=3.8.0
requests>=2.31.0
tenacity>=8.0.0

# Office tools
python-docx>=0.8.11
openpyxl>=3.1.2
python-pptx>=0.6.21
pywin32>=306; sys_platform == 'win32'

# Web and UI
flask>=3.0.0
streamlit>=1.32.0
streamlit-chat>=0.1.1
streamlit-extras>=0.3.5
beautifulsoup4>=4.12.0
duckduckgo-search>=4.1.1

# Data processing
pandas>=2.0.0
numpy>=1.23.0
matplotlib>=3.5.0
seaborn>=0.12.0

# Text processing
textblob>=0.17.0
spacy>=3.5.0
nltk>=3.8.0

# Development tools
pylint>=2.17.0
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0
pytest>=7.3.0
coverage>=7.2.0

# System monitoring
psutil>=5.9.0
prometheus-client>=0.17.0
redis>=4.5.0

# Security
bcrypt>=4.0.0
cryptography>=40.0.0

# Documentation
sphinx>=6.1.0
docutils>=0.19.0

# MCP Server
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
aiofiles>=23.1.0